+++ src/restfw/root.py
@@ -3,6 +3,7 @@
 :Authors: cykooz
 :Date: 20.08.2016
 """
+from pyramid.registry import Registry
 from pyramid.security import Allow, DENY_ALL, Everyone
 from zope.interface import implementer
 
@@ -15,16 +16,12 @@ from .utils import notify
 
 @implementer(IRoot)
 class Root(SimpleContainer):
-
     __acl__ = [
         (Allow, Everyone, ALL_GET_REQUESTS),
         DENY_ALL
     ]
 
-    def __init__(self, registry):
-        """
-        :type registry: pyramid.registry.Registry
-        """
+    def __init__(self, registry: Registry):
         super(Root, self).__init__()
         self.registry = registry